Check for function

+++ inc/functions/header-charm.php
@@ -5,6 +5,8 @@
  * Show a newsletter signup in the header.
  */
 add_action( 'exa_header_charm', function( $content ) {
-	echo "<h3>Weekly Newsletter</h3>";
-	echo mc4wp_get_form();
+	if( function_exists('mc4wp_get_form') ) {
+		echo "<h3>Weekly Newsletter</h3>";
+		echo mc4wp_get_form();
+	}
 });
